Add tests for `BushAuthoritySystem`
Also, fix a previously-undiscovered (but possibly minor) bug in which
it appears that the bush entity backing an unsustainable bush was not
destroyed on harvest, even though the block was set to air.  It is not
clear whether this was actually a problem or not, but it clearly does
no harm to destroy the entity in question.